some of the spin-apiserver done now

+++ design/types.go
@@ -6,7 +6,7 @@ import . "goa.design/goa/v3/dsl" // nolint
 var Storage = Type("Storage", func() {
 	Attribute("volume", String, "Volume name, must not include `/`")
 	Attribute("image", String, "Image filename, must not include `/`")
-	Attribute("image_size", UInt, "Image size (in gigabytes)")
+	Attribute("image_size", UInt64, "Image size (in gigabytes)")
 	Attribute("cdrom", Boolean, "Is this image a cdrom?")
 	Required("volume", "image")
 })
